+++ assets/app.js
@@ -1,2 +1,3 @@
 import './stylesheets/application.scss';
 import './javascripts/application.js';
+import './javascripts/cookie.js';