+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Implement google analytics alongside plausible
+
 ### Changed
 
 ## [release-036] - 15/05/2024